+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## v1.1.0
+
+- Breaking change - The access token is a short-lived token - Authentication is now done via OAUTH2.0 client credentials flow
+  - Removed support for setting an access token via `STEADYBIT_EXTENSION_ACCESS_TOKEN` or `appdynamics.accessToken`
+  - Added parameters client name, client secret and account name to the configuration.
+
 ## v1.0.0
 
  - Initial release